Minimize Cognitive Complexity of edit method
